Change registers to hold 64 bit values
This directly supports doubles as floating point registers would need
to be able to hold 64 bit values. It also should make it easier to
implement RV64 in the future should that ever be done.

+++ src/rars/Launch.java
@@ -573,7 +573,7 @@ public class Launch {
             } else { // Integer register or CSR
                 if (verbose)
                     out.print(reg + "\t");
-                out.println(formatIntForDisplay(RegisterFile.getRegister(reg).getValue()));
+                out.println(formatIntForDisplay((int)RegisterFile.getRegister(reg).getValue()));
             }
         }
     }
